Update leaflet.timedimension.layer.wms.js
Using GeoWebCache, the GetCapabilities method does not contains the time dimension. I need to access the original WMS server to retrieve the correct data. Also, using GeoWebCache, the layer name can be different from the layer on the original WMS server, so I need also the layer name on the original WMS server.

Added the "getCapabilitiesUrl" to specify an alternate url for the GetCapabilities request

Added the "getCapabilitiesLayerName" to specify an alternate layer name for the GetCapabilities request