Support for display WMS layers over a continuous interval
Added the possibility to display a WMS layers not for a single time but
for a startTime/endTime continuos interval if supported by the WMS
server (like Geoserver).

-leaflet.timedimension.control-
added a _setDisplayDateValue used to display startTime/endTime on
displayDate element

-leaflet.timedimension.js-
added the fixedStartTime options. If true, the first available time is
used as startTime

-leaflet.timedimension.layer.wms.js-
modified the _createLayerForTime function to create the
startTime/endTime time parameter if enabled.

Added an hack for the _baselayer creation in the
_requestTimeDimensionFromCapabilities function (without this hack the
base layers time parameter contain only the default time and not the
startTime/endTime value). This need a better solution.  *1

Since sometimes the GetCapabilities request can be slow, I've added
"this._timeDimension.fire('timeloading', {});"  to the _update funtion
so the user can see some visual feedback. This need a better solution.
*2

Derived from my previous Pull request (#71) I've modified the isLoaded
function to check for boundingbox intersection between the map and the
layers; if a layer is outside the boundingbox of the map, isLoaded is
forced to true (otherwise isLoaded is always false).

*1 and *2 need some love, all suggestions are appreciated ^_^