Fix first frame skipped in startover (refs #184)
Start method from the player had a bug when starting again
after an animation has finished (with time at max index and the startOver
option activated). Time was set to 0 and immediately the first tick was being
called, so time was moving to the second time index. Now, in this case,
we prevent the first tick to be called.