Fix issue with entire GeoJSON being displayed at last feature time
When a last feature time is equal to the current min_time, the current
logic leaves min_index set to null. This then sets the min_time to the
zero index, and such the entire layer is erroneously displayed.

This change simply changes to check the feature time is greater than
**or equal to** to the min time.